Shadows and light optimizations Added a simple system to figure out whether some pixels is in shadow or not by casting a ray at different lights and seeing if the ray intersects anything. This provides some quite good results for point lights, but it butchered the performance from 120-180 fps to 15-30fps.
Becuase of that, now I'm importing the max range of a point light from
model file if that value exists. If it doesn't, I set it to 10 000. This
way I can completely ignore lights that are far enough away, improving
the performance to 55-70fps.
I think the base version of the ray tracing system is now done. It's
capable of doing everything that rasteriser can even do shadows.
Performance isn't too terrible for the version either as long as the
amount of lights isn't too high.