Performance and looks Applied resampled importance sampling to the closest hit shader for selecting the reservoir. You know, the algorithm that pretty much all of my code meant for improving many lights performance is centered around? This way I only need to go through a small amount of reservoirs and still get a good result. Currently, I have 128 reservoirs per cell, but I'm only grabbing 8 samples in closest hit shader. That sounds like it should give a massive speed boost, but it's not quite so simple. In order to get good results, I need to be able to grab a good sample randomly. In reservoir grid constructor this is simple, every light has the same chance. But in closest hit shader, different lights have different chances, meaning I have to do weighted randomness instead of uniform, making this a decent bit harder.
I solved it by precalculating a so called cumulative distribution
function of all the reservoirs propability density functions for all
cells in a compute shaders, which only takes like 0,3 milliseconds
despite being kinda unoptimal. Then in the closest hit shader I use
binary search to find the reservoir that is in the correct place in cdf
for the uniformly random float. Kinda hard to explain quickly, but this
does the weighted randomness 100% correctly, but this is kinda slow as
it involves a loop, bunch of if statements and incoherent memory
accesses.
Still, it boosts my fps by on average 2x.
This system also prioritizes more contributing lights a lot more than
old system, meaning the sunlight and lamps are a lot more clearly
visible now due to being strong where as weaker candles are only
noticable in few different places. This looks a bit nices in my opinion.
Tho I think it prefers strong lights a little too much, as I think due
to candles large amount their influence should be a bit larger.