Integrated the async image loading into vulkano Basically just took the async image loading code I had in the raytracers main file and slapped it onto the GltfLoader. I made a new function for that as I still want to keep the sync image loader. Dealing with this image stuff was getting a little annoying with the scene, so I decided to move to not import images in the scene code, but instead return the GltfLoader so I can more easily decide how I want to load the images.
Unfortunately, not all things are well. For whatever reason,
synchronizing between threads is harder when the thread comes from
another class instead of being part of the main function. This lead to
me using overly cautios synchronization to get correct results. And by
that I mean wrapping the entire image loading code in a mutex so I can't
access the images when the thread is doing anything at all with them. I
also had to put a giant 50ms delay between loading each image, because
if that delay was as "small" as 20ms, some images would never get fully
loaded, which is weird. This slows down the image loading a bit. Also,
because the entire image loading code is wrapped in a mutex and loading
an image can take nearly 100ms, whenever I would in the main render loop
lock the mutex so I could update the descriptors, there would be a very
big lag spike. Also, and this was the problem with the async image
loading in the main function too, I'm only using a single thread for the
ktx image transcoding to make things easier and so loading all the
images takes a while. Also, I don't really like the changes I made to
the Scene class to make loading images easier, so I think I have to
redesign how I load gltf files in to the Scene.
So while the async image loading is technically working, it certainly is
in need of further refinements.