Progress on video and better support for my laptop On the VulDevice I added ability to enable video decoding, which will enable a few extensions needed for video decoding and will get the index of the video decoding capable queue family. I also now create a video session, get it's memory requirements, create a few chunks of memory as dictated by the memory requirements and bind them to the session. Important to note these aren't some fancy buffers or even fancier multiplanar images that I am scared of, they are just raw chunks of memory needed to hold some of the video sessions internal data.
This commit was done on my laptop and not on my desktop like usual, and
I noticed a little issue with VulDevice. It just iterates through every
physical device and picks the first one that has support for all the
needed features. It was fine on my desktop since it only got 1 gpu, but
my laptop got a crappy integrated gpu and a slighly less crappy discrete
gpu. And worst of all, the integrated gpu appears first on the list. So
I just wrote a simple priority system which prioritises certain types of
physical devices over others. This way the discrete gpu will always be
used if it is available and supports all necessary features.